fix: refactor deprecation renderer, show request/notification deprecations
